Improve the systemd custom dependencies
The mounts no longer depend on ocfs2.service (that was silly)

nfs-server.service no longer depends on the mounts. Instead
nfs-mountd.service now depends on ocfs2.service. The first
depends on the later and this fails to start without
/var/lib/nfs.

Last, keepalived.service depends on nfs-server.service instead
of the mounts.

So now we have

ocfs2 (mounts) > nfs-mountd > nfs-server > keepalived